requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
python-multipart
Pillow
numpy
opencv-python
